E2E tests for rspec-ruby with alias rspec command
