+++ features/rspec-ruby/bin/retest
@@ -1,0 +1,3 @@
+#!/usr/bin/env sh
+
+bundle exec rspec $@
