--- features/rspec-ruby/retest/retest_test/setup_test.rb
@@ -1,5 +1,0 @@
-class SetupTest < Minitest::Test
-  def test_repository_setup
-    assert_equal :rspec, Retest::Setup.new.type
-  end
-end
